# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/de/fhe/ai/colivingpilot/MainActivity.kt
#	app/src/main/java/de/fhe/ai/colivingpilot/core/CoLiPiApplication.kt
#	app/src/main/java/de/fhe/ai/colivingpilot/settings/WgViewmodel.kt
#	app/src/main/java/de/fhe/ai/colivingpilot/storage/Repository.kt
#	app/src/main/java/de/fhe/ai/colivingpilot/storage/UserDao.kt
#	app/src/main/res/navigation/nav_graph.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/de/fhe/ai/colivingpilot/storage/Repository.kt
@@ -20,21 +20,79 @@ class Repository(
     private val shoppingListItemDao: ShoppingListItemDao = db.shoppingListItemDao()
     private val taskAssignedUserDao: TaskAssignedUserDao = db.taskAssignedUserDao()
 
-    suspend fun updateAll() = withContext(Dispatchers.IO) {
-        userDao.deleteAll() // triggers cascading deletes in all tables
-        // TODO: ... get json data from backend and insert everything
+    /**
+     * Refreshes the app's data by fetching the relevant dataset from the server.
+     *
+     * @return A pair indicating the success or failure of the operation and an optional error message.
+     */
+    suspend fun refresh(): Pair<Boolean, String> = withContext(Dispatchers.IO) {
+        try {
+            val response = RetrofitClient.instance.getWgData().execute()
+            if (!response.isSuccessful) {
+                val body = response.errorBody()
+                body?.let {
+                    val json = JSONObject(body.string())
+                    return@withContext Pair(false, json.getString("status"))
+                }
+                return@withContext Pair(false, "")
+            }
+
+            CoLiPiApplication.instance.keyValueStore.writeString("wg_name", "")
+            CoLiPiApplication.instance.keyValueStore.writeString("wg_code", "")
+            CoLiPiApplication.instance.keyValueStore.writeInt("wg_max_members", -1)
+
+            userDao.deleteAll() // triggers cascading deletes in all tables
+            taskDao.deleteAll() // TODO: removable once Task has reference to User
+
+            val body = response.body()
+            body?.let { resp ->
+                CoLiPiApplication.instance.keyValueStore.writeString("wg_name", resp.data.wg.name)
+                CoLiPiApplication.instance.keyValueStore.writeString("wg_code", resp.data.wg.invitationCode)
+                CoLiPiApplication.instance.keyValueStore.writeInt("wg_max_members", resp.data.wg.maximumMembers)
+                val creator = resp.data.wg.creator
+                resp.data.wg.members.forEach { member ->
+                    db.userDao().insert(User(member.id, member.username, member.beercounter, creator.username == member.username))
+                }
+                resp.data.wg.shoppingList.forEach { item ->
+                    db.shoppingListItemDao().insert(ShoppingListItem(item.id, item.title, item.notes, item.creator.id, item.isChecked))
+                }
+                resp.data.wg.tasks.forEach { task ->
+                    db.taskDao().upsert(Task(task.id, task.title, task.description, task.beerbonus))
+                }
+            }
+            return@withContext Pair(true, "")
+        } catch (_: Exception) {
+            Log.e(CoLiPiApplication.LOG_TAG, "Failed to fetch WG data")
+            return@withContext Pair(false, "")
+        }
     }
 
     fun getUsersFlow(): Flow<List<User>> {
         return userDao.getUsersFlow()
     }
 
-    fun getTasks(): List<Task> {
+    fun addOrUpdateTask(task: Task) {
+        return taskDao.upsert(task)
+    }
+
+    fun deleteTask(task: Task) {
+        return taskDao.delete(task)
+    }
+
+    fun deleteTaskById(id: String) {
+        return taskDao.deleteByID(id)
+    }
+
+    fun getTasks(): Flow<List<Task>> {
         return taskDao.getTasks()
     }
 
-    fun getShoppingListItems(): List<ShoppingListItem> {
-        return shoppingListItemDao.getShoppingListItems()
+    fun getTask(id: String): Flow<Task> {
+        return taskDao.getTask(id)
+    }
+
+    fun getShoppingListItemsFlow(): Flow<List<ShoppingListItem>> {
+        return shoppingListItemDao.getShoppingListItemsFlow()
     }
 
     suspend fun addUser(user: User) {
@@ -50,4 +108,21 @@ class Repository(
     }
 
 
+    fun deleteItemFromShoppingList(shoppingListItem: ShoppingListItem){
+        shoppingListItemDao.deleteItemFromShoppingList(shoppingListItem)
+    }
+
+    fun insertShoppingListItem(shoppingListItem: ShoppingListItem){
+        shoppingListItemDao.insert(shoppingListItem)
+    }
+
+    fun updateItem(shoppingListItem: ShoppingListItem, boolean: Boolean){
+        var updatedItem = shoppingListItem.copy(isChecked = boolean)
+        shoppingListItemDao.updateItem(updatedItem)
+    }
+
+    //TODO später kommt der User aus der Anmeldung
+    fun getTestUser(): User {
+        return userDao.getTestUser()
+    }
 }
